Extract Feed Image Loader Spy into separate file to avoid duplication
